1. Fixed broken build-and-serve-debug script. 1. Code refactoring. 1. Cool stuff and bug fixes from the new ALS update.
