+++ SynthLineLayer.js
@@ -1,0 +1,10 @@
+L.ALS.SynthLineLayer = L.ALS.SynthBaseDrawLayer.extend({
+	defaultName: "Line Layer",
+	drawControls: {
+		polyline: {
+			shapeOptions: {
+				color: "#ff0000"
+			}
+		}
+	},
+});
