+++ SynthPolygonLayer.js
@@ -1,0 +1,10 @@
+L.ALS.SynthPolygonLayer = L.ALS.SynthBaseDrawLayer.extend({
+	defaultName: "Polygon Layer",
+	drawControls: {
+		polygon: {
+			shapeOptions: {
+				color: "#ff0000"
+			}
+		}
+	},
+});
