+++ index.html
@@ -16,9 +16,11 @@
 	<!-- Leaflet.Coordinates -->
 	<link rel="stylesheet" href="node_modules/leaflet.coordinates/dist/Leaflet.Coordinates-0.1.5.css" />
 
-	<!-- ALS stylesheets -->
+	<!-- Leaflet.Draw -->
+	<link rel="stylesheet" href="node_modules/leaflet-draw/dist/leaflet.draw.css" />
+
+	<!-- ALS stylesheet -->
 	<link rel="stylesheet" href="node_modules/leaflet-advanced-layer-system/dist/css/base.css" />
-	<link rel="stylesheet" href="node_modules/leaflet-advanced-layer-system/dist/css/dark.css" id="als-dark-theme-link" />
 
 	<!-- SynthFlight stylesheet -->
 	<link rel="stylesheet" href="css/styles.css" />
