+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "synthflight",
   "productName": "SynthFlight",
-  "version": "0.0.12-alpha",
+  "version": "0.0.13-alpha",
   "description": "A fully client-side software for planning aerial photography",
   "main": "electronApp.js",
   "browser": "index.html",
@@ -45,7 +45,7 @@
     "serve": "http-server \"./dist/SynthFlight-browser\" -c-1",
     "build": "node build.js",
     "build-and-serve": "node build.js -b && npm run serve",
-    "build-and-serve-debug": "node build.js -b -d && npm run build-and-serve"
+    "build-and-serve-debug": "node build.js -b -d && npm run serve"
   },
   "keywords": [],
   "author": "matafokka",
@@ -80,7 +80,8 @@
     "http-server": "^0.12.3",
     "keyboardevent-key-polyfill": "^1.1.0",
     "leaflet": "^1.7.1",
-    "leaflet-advanced-layer-system": "^1.2.0",
+    "leaflet-advanced-layer-system": "file:../leaflet-advanced-layer-system",
+    "leaflet-draw": "^1.0.4",
     "leaflet.coordinates": "~0.1.5",
     "persistify": "^2.0.1",
     "postcss": "^8.2.4",
